+++ Dragoon Modifier/MemoryController/Shop.cs
@@ -6,13 +6,13 @@ using System.Threading.Tasks;
 
 namespace Dragoon_Modifier.MemoryController {
     public class Shop {
-        long _baseAddress;
+        int _baseAddress;
         ByteCollection _item;
 
         public byte WeaponItemFlag { get { return Emulator.ReadByte(_baseAddress); } set { Emulator.WriteByte(_baseAddress, value); } }
         public ByteCollection Item { get { return _item; } }
 
-        public Shop(long baseAddress, int shop) {
+        public Shop(int baseAddress, int shop) {
             _baseAddress = baseAddress + shop * 0x40;
             _item = new ByteCollection(_baseAddress + 1, 4, 16);
